check_bugowner: Really fix missing directories.

+++ check_bugowner.py
@@ -140,7 +140,7 @@ class CheckerBugowner(ReviewBot.ReviewBot):
         local_dir = Path(self._gitea_cache_dir(), owner, repo)
         self.logger.debug(f"Cache directory: {local_dir}")
 
-        if not self.scm.is_repo(local_dir):
+        if not self.scm.is_repo(local_dir) and local_dir.exists():
             self.logger.warning(f"Git cache directory appears corrupted, removing {local_dir}...")
             shutil.rmtree(local_dir)
